+++ R/console.R
@@ -26,6 +26,13 @@ julia_complete_console <- function(){
     message("Preparing julia REPL, press Ctrl + D to quit julia REPL.")
     message("You could get more information in how to use julia REPL at <https://docs.julialang.org/en/stable/manual/interacting-with-julia/>")
     julia_command("Base._start()")
+
+    if (julia$useRCall) {
+        ## in the hook RCall.rgui_start() will be executed,
+        ## then when we quit the console,
+        ## RCall.rgui_stop() needs to be executed.
+        julia_command("RCall.rgui_stop()")
+    }
 }
 
 julia_incomplete_console <- function(){